♻️ refactor(starlite): support adding background tasks
